Begzod

+++ scr/MVC Producr/MVC Producr/Controllers/AdminController.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using MVC_Producr.Data;
-using MVC_Producr.Models;
 using MVC_Producr.Models.Domain.Entities;
+using MVC_Producr.Models.ViewModels;
 
 namespace MVC_Producr.Controllers
 {
@@ -20,9 +20,27 @@ namespace MVC_Producr.Controllers
         {
             return View();
         }
+        public IActionResult Post()
+        {
+            return View();
+        }
+        public async Task<IActionResult> Get()
+        {
+            var products = await _context.Products.ToListAsync();
+            return View(products);
+        }
+        public IActionResult Delete()
+        {
+            return View();
+        }
+
+        public IActionResult Update()
+        {
+            return View();
+        }
 
         [HttpPost]
-        public async Task<IActionResult> CreateProduct(ProductViewModel product)
+        public async Task<IActionResult> Add_Product(ProductViewModel product)
         {
 
             var _product = new Product()
@@ -34,41 +52,38 @@ namespace MVC_Producr.Controllers
 
             await _context.Products.AddAsync(_product);
             await _context.SaveChangesAsync();
-            return View(product);
+            return RedirectToAction("Index","Admin");
         }
 
 
-        [HttpGet]
-        public async Task<IActionResult> Get()
+        [HttpPost,ActionName("Delete")]
+        public async Task<IActionResult> Delete_Product(DeleteViewModel ProductName)
         {
-            var products = await _context.Products.ToListAsync();
-            return View(products);
-        }
-
-
-        [HttpDelete]
-        public async Task<IActionResult> Delete(string ProductName)
-        {
-            var product = await _context.Products.FirstOrDefaultAsync(x => x.ItemName == ProductName);
+            var product = await _context.Products.FirstOrDefaultAsync(x => x.Id == ProductName.Id);
+            if (product == null)
+            {
+                return RedirectToAction("Index","Home");
+            }
 
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
+            return RedirectToAction("Index", "Admin");
 
-            return View(product);
         }
 
 
-        [HttpPut]
-        public async Task<IActionResult> Update(string product,ProductViewModel productChanges)
+        [HttpPost,ActionName("Update")]
+        public async Task<IActionResult> Update_Product(UpdateViewModel model)
         {
-            var productData = await _context.Products.FirstOrDefaultAsync(x=>x.ItemName == product);
+            var productData = await _context.Products.FirstOrDefaultAsync(x=>x.Id == model.Id);
 
-            productData.ItemName = productChanges.ItemName;
-            productData.Quantity = productChanges.Quantity;
-            productData.Price = productChanges.Price;
+            productData.ItemName = model.ItemName;
+            productData.Quantity = model.Quantity;
+            productData.Price = model.Price;
 
             await _context.SaveChangesAsync();
-            return View(productData);
+            return RedirectToAction("Index", "Admin");
+
         }
 
     }
